+++ code/go/sdk/consts.go
@@ -5,7 +5,7 @@ import "time"
 const (
     Version                   = "0.20.0-beta"
     VersionClientByteSize     = 43609
-    VersionClientByteSizeGzip = 14889
+    VersionClientByteSizeGzip = 14886
     DatastarKey               = "datastar"
     DefaultSettleTime         = 300 * time.Millisecond
     DefaultSSERetryDuration   = 1000 * time.Millisecond