+++ code/ts/library/src/plugins/preprocessors/core/actions.ts
@@ -1,11 +1,16 @@
 import { PreprocessorPlugin, RegexpGroups } from "../../../engine";
 import { wholePrefixSuffix } from "../../../utils/regex";
 
-// Replacing $$action(args) with ctx.actions.action(ctx, args)
+// Replacing @action(args) with ctx.actions.action(ctx, args)
 export const ActionsProcessor: PreprocessorPlugin = {
     name: "action",
     pluginType: "preprocessor",
-    regexp: wholePrefixSuffix("$\\$", "action", "(?<call>\\((?<args>.*)\\))?"),
+    regexp: wholePrefixSuffix(
+        "@",
+        "action",
+        "(?<call>\\((?<args>.*)\\))?",
+        false,
+    ),
     replacer: ({ action, args }: RegexpGroups) => {
         const withCtx = [`ctx`];
         if (args) {